lisää oppiaine opetussuunnitelmaan

+++ eperusteet-ylops-service/src/main/java/fi/vm/sade/eperusteet/ylops/service/impl/TekstiKappaleViiteServiceImpl.java
@@ -140,7 +140,8 @@ public class TekstiKappaleViiteServiceImpl implements TekstiKappaleViiteService 
         }
 
         if (viite.getTekstiKappale() != null &&
- viite.getTekstiKappale().getTila().equals(Tila.LUONNOS) &&            findViitteet(opsId, viiteId).size() == 1) {
+            viite.getTekstiKappale().getTila().equals(Tila.LUONNOS) &&
+            findViitteet(opsId, viiteId).size() == 1) {
             TekstiKappale tekstiKappale = viite.getTekstiKappale();
             tekstiKappaleService.delete(tekstiKappale.getId());
         }
